Split bundle/prune.rs into focused submodules
prune/mod.rs keeps the prune entry point, candidate reporting, and the
suggested-command helper; assess.rs owns per-bundle/per-repo dead-work
signals and the shared PR/checkout scan cache; orphans.rs owns orphan
worktree discovery/removal and KnitHub remote orphan classification.
Pure moves, no behavior change.

Knit-Group: kg_20260611_f9bbb1
Knit-Bundle: open-source-prep
